requirements.txt
asgiref==3.8.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.3
django-admin==2.0.2
django-excel-response2==3.0.6
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
idna==3.6
isoweek==1.3.3
jwcrypto==1.5.6
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.31.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
xlwt==1.3.0
